clippy-backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
python-dotenv>=1.0.0
qdrant-client>=1.16.0
llama-cpp-python>=0.3.0
cognee>=0.5.2
cognee-community-vector-adapter-qdrant>=0.2.0
pydantic>=2.0
requests>=2.32.0
